\name{dict}
\alias{dict}
\docType{data}
\title{
University Name Dictionary
}
\description{
The dict dataset includes the parsed names of many popular statistics graduate
programs. It is used to standardize the graduate program names, enabling 
aggregate analysis. 
}
\usage{data("dict")}
\format{
  The format is:
 chr [1:111, 1] "arizona state univ" "auburn univ" "baylor univ" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr "name"
}
\examples{
data(dict)

}
\keyword{datasets}
